Upgrade to latest rust (DANGEROUS)
Closes https://github.com/bjz/gl-rs/issues/76

Notice the transmute is replaced by direct casting by pointers. I do
not know how this affects the usage of the library. I left the ‘unused
import’ warning on in case this can be fixed later. If this is
critical, I would recommend doing an analysis and document which uses
cases one should be careful.