Fix missing dyn warning
The latest Rust nightly will print a warning when the `dyn` keyword is
not present. This adds the keyword to prevent the warning from popping
up downstream.

An alternative would be to add an annotation to ignore the warning,
which would support more outdated Rust versions, however all recent Rust
versions should support the `dyn` keyword.